Converted existing Super Editor keyboard interactor into a generically named widget and used it in the IME interactor
