+++ super_editor/lib/src/default_editor/document_hardware_keyboard/document_input_keyboard.dart
@@ -1,0 +1,10 @@
+export 'document_physical_keyboard.dart';
+export 'document_keyboard_actions.dart';
+
+/// This file exports various document hardware keyboard tools.
+///
+/// A document editor might be configured to work exclusively with
+/// hardware keyboard keys. Or, what's more likely, is that a document
+/// editor is configured to process Input Method Engine (IME) input
+/// for most content editing, but the editor still receives and processes
+/// hardware key events for things like arrow keys, tab keys, etc.
