+++ super_editor/lib/super_editor.dart
@@ -24,8 +24,7 @@ export 'src/infrastructure/document_gestures.dart';
 export 'src/default_editor/document_gestures_mouse.dart';
 export 'src/default_editor/document_gestures_touch.dart';
 export 'src/default_editor/document_ime/document_input_ime.dart';
-export 'src/default_editor/document_input_keyboard.dart';
-export 'src/default_editor/document_keyboard_actions.dart';
+export 'src/default_editor/document_hardware_keyboard/document_input_keyboard.dart';
 export 'src/default_editor/horizontal_rule.dart';
 export 'src/default_editor/image.dart';
 export 'src/default_editor/layout_single_column/layout_single_column.dart';
